Post to variable slack channels
If you define the `SiteSetting.slack_channel` to include `{category}`,
the plugin will substitute the `Post` category name.

For example:

If you have defined:

`SiteSetting.slack_channel = "#discourse-{category}"`

and you create a new Post in with a Category of "Servers", the
notification will get posted to the `#discourse-servers` channel.