feat: Category specific channels
This adds the ability to turn on category specific channels, which
can be specified in the category settings when the setting is enabled.

If the category does not have a specific channel set, it falls back to
the parent category if there is one, otherwise it just uses the global
channel specified in the settings.